requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
insightface==0.7.3
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3
scikit-learn==1.3.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
